refactor: divide responsibilities over handlers in preparation for unit tests
